Separated Flashcard code from learn component
